Update componets and moved firebase connection to Home.vue
